Move save_all into serialize module.
The sherpa.astro.ui.utils module is large enough, so the two
routines involved with state serialization (save_all and save_session)
have been moved out to a separate module (sherpa.astro.ui.serialize).
